+++ pennylane/transforms/__init__.py
@@ -83,6 +83,7 @@ both transforms, and decompositions within the larger PennyLane codebase.
     ~transforms.xyx_decomposition
     ~transforms.two_qubit_decomposition
     ~transforms.set_decomposition
+    ~transforms.ctrl_decomp_zyz
     ~transforms.pattern_matching
     ~transforms.to_zx
     ~transforms.from_zx
@@ -195,7 +196,12 @@ from .batch_partial import batch_partial
 from .classical_jacobian import classical_jacobian
 from .condition import cond, Conditional
 from .compile import compile
-from .decompositions import zyz_decomposition, xyx_decomposition, two_qubit_decomposition
+from .decompositions import (
+    zyz_decomposition,
+    xyx_decomposition,
+    two_qubit_decomposition,
+    ctrl_decomp_zyz,
+)
 from .defer_measurements import defer_measurements
 from .hamiltonian_expand import hamiltonian_expand, sum_expand
 from .split_non_commuting import split_non_commuting
